Made terrain analysis pick better starting points and increased the distance check in economy to match the changed distance measure using distance to dropsite.
Terrain analysis now only picks a start point if it is within a region
of size >10.
The economy code only does a distance check if there are >100 entities
of that resource on the map so rare resources are gathered.
Modified some debug output as well.

This was SVN commit r10773.