Fix an issue that caused a crash on serialization. Prepare the AI manager a little more for serialization.
Fix bugs with tests. Fix some other issues in the AI (attack plans deal
with walls better, choose better paths, target selection is better. Dock
building won't be tried hundreds of times if it fails.)
Changes the Oasis random map to only put a path in the middle 50% of the
time.

This was SVN commit r13230.