[PetraAI] - Teach PetraAI to fear fire.
In 86ddf09640, fire damage was introduced for the Iberian fire ship (and
the Theban fire raiser), this adds that damage type to the config of
PetraAI.
Poison (a24c8419c5) is only used in Status Effects (which PetraAI
doesn't understand anyway) so is not added.

Differential revision: D3432
Reviewed by: @Angen
This was SVN commit r24758.